+++ src/Composer/Downloader/HgDownloader.php
@@ -25,6 +25,8 @@ class HgDownloader extends VcsDownloader
      */
     public function doDownload(PackageInterface $package, $path, $url)
     {
+        $this->checkSecureHttp($url);
+
         $url = ProcessExecutor::escape($url);
         $ref = ProcessExecutor::escape($package->getSourceReference());
         $this->io->writeError("    Cloning ".$package->getSourceReference());
@@ -43,6 +45,8 @@ class HgDownloader extends VcsDownloader
      */
     public function doUpdate(PackageInterface $initial, PackageInterface $target, $path, $url)
     {
+        $this->checkSecureHttp($url);
+
         $url = ProcessExecutor::escape($url);
         $ref = ProcessExecutor::escape($target->getSourceReference());
         $this->io->writeError("    Updating to ".$target->getSourceReference());
@@ -83,6 +87,13 @@ class HgDownloader extends VcsDownloader
         }
 
         return $output;
+    }
+
+    protected function checkSecureHttp($url)
+    {
+        if (preg_match('{^http:}i', $url) && $this->config->get('secure-http')) {
+            throw new TransportException("Your configuration does not allow connection to $url. See https://getcomposer.org/doc/06-config.md#secure-http for details.");
+        }
     }
 
     /**